Move RequestAndCallback from PaxosManager to PaxosConfig
